requirements.txt

bigchaindb_driver==0.5.0a4
eth-brownie==1.18.1
python-dotenv==0.16.0